Automated Dive Timer (Issue #9)
The dive timer now automatically starts when the ROV enters the water,
pauses when the ROV is at the surface and then automatically resumes if
the ROV re-enters the water.  It is only reset if the user chooses to
reset it from the UI.  It has been moved to its own class, DiveTimer.
Also, the clock in the UI has been changed from a 12 hour clock with an
AM/PM marker to a 24 hour clock.