Misc. Polishing
Remove "beta" from application title.
Fix bug in LedIndicator's C++ class (a stupid bug, at that).
Fix bug where indicators would not light up if status should have been
set to "on" when application started.  For instance, if the RPi was
already sending data when the application started, it would not display
the RPi as being connected because the signal/slot wasn't connected by
that time.