Started adding gears support and disabled state
The ROV is "disabled" when in gear 0.  At that point, Monterey simply stops
sending packets.  The ROV should then go into a lost com state, disabling
everything.

The gears are configurable by the user.  Right now, the '-' and '=' keys
decrease and increase the gearing, respectively.  I plan on adding a UI element
and joystick support for that, but I think I should add a custom QEvent setup
for that.  I need to look into how I want to handle that.